move prisma service in common module
