+++ src/common/common.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { LoggerService } from './logger/logger.service';
+import { PrismaService } from './prisma/prisma.service';
 
 @Module({
-  providers: [LoggerService],
-  exports: [LoggerService],
+  providers: [LoggerService, PrismaService],
+  exports: [LoggerService, PrismaService],
 })
 export class CommonModule {}
