+++ src/common/prisma/prisma.service.ts
@@ -1,0 +1,43 @@
+import { Injectable, OnModuleInit } from '@nestjs/common';
+import { Prisma, PrismaClient } from '@prisma/client';
+import { LoggerService } from '../logger/logger.service';
+
+@Injectable()
+export class PrismaService
+  extends PrismaClient<Prisma.PrismaClientOptions, Prisma.LogLevel>
+  implements OnModuleInit
+{
+  constructor(private readonly logger: LoggerService) {
+    super({
+      log: [
+        { emit: 'event', level: 'query' },
+        { emit: 'event', level: 'info' },
+        { emit: 'event', level: 'warn' },
+        { emit: 'event', level: 'error' },
+      ],
+    });
+  }
+
+  async onModuleInit() {
+    this.$on('query', (event) => {
+      this.logger.debug(event.query, {
+        params: event.params,
+        duration: `${event.duration} ms`,
+      });
+    });
+
+    this.$on('info', (event) => {
+      this.logger.log(event.message);
+    });
+
+    this.$on('warn', (event) => {
+      this.logger.warn(event.message);
+    });
+
+    this.$on('error', (event) => {
+      this.logger.error(event.message);
+    });
+
+    await this.$connect();
+  }
+}
