--- src/prisma/prisma.service.ts
@@ -1,25 +1,0 @@
-import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
-import { Prisma, PrismaClient } from '@prisma/client';
-
-@Injectable()
-export class PrismaService
-  extends PrismaClient<Prisma.PrismaClientOptions, Prisma.LogLevel>
-  implements OnModuleInit
-{
-  private readonly logger = new Logger(PrismaService.name);
-  constructor() {
-    super({ log: ['query', 'info', 'warn', 'error'] });
-  }
-
-  async onModuleInit() {
-    this.$on('query', (event) => {
-      this.logger.log(
-        `Query: ${event.query}`,
-        `Params: ${event.params}`,
-        `Duration: ${event.duration} ms`,
-      );
-    });
-
-    await this.$connect();
-  }
-}
